mod Separate gpkg import code
will be easier to test this way
